Apply @ihhub suggestions

+++ src/fheroes2/gui/ui_tool.h
@@ -110,7 +110,7 @@ namespace fheroes2
     };
 
     // Fade display image colors to grayscale part of default game palette.
-    void colorFade( const std::vector<uint8_t> & palette, const fheroes2::Rect & frameRoi, const uint32_t durationMs, const uint32_t fps );
+    void colorFade( const std::vector<uint8_t> & palette, const fheroes2::Rect & frameRoi, const uint32_t durationMs, const double fps );
 
     void CreateDeathWaveEffect( Image & out, const Image & in, const int32_t x, const std::vector<int32_t> & deathWaveCurve );
